Add basic RUN_COMMAND message handling
Instead of declaring msg, msg_type, and msg_size in ipc_read_client, declare
the variables in dwm.c, so they can further be processed for further message
handling.

ipc_parse_run_command uses yajl to parse the client's command and arguments and
then uses the Arg structure as defined in dwm.c to store each of the arguments
which can then be used to call the apporiate functions in dwm.c

The commands are also mapped into enum values for the sake of enumerating the
available commands as well as avoiding passing around allocated strings.