Move Arg struct to separate types.h file
Some types must be known by the ipc module as well as dwm.c. The only way to
have these files share the same type definitions is to pull them into a separate
header file to be included. Unfortunely, this does impact the compatability of
this IPC patch since it results in more modifications to the main dwm.c file.