Implement IPC message sending to clients
Add ipc_prepare_send_message to append the message to be sent to the specified
IPC client's buffer. This function does not start writing the message to the
client's socket right away in case the client is not ready.

Add ipc_push_pending to attempt to write the pending buffer of a specified
client to the client's socket. If the entire buffer cannot be written, the
buffer is shifted to the amount left to write and reallocated.

Add ipc_write_message to attempt to write a given buffer to a specified file
descriptor. This function will return with an error if any part of the buffer
could not be written, otherwise it returns the amount of bytes written.

Since the IPC magic string does not contain a null character, the IPC_MAGIC_ARR
is defined as a char array of the string without the null char. This simplifies
initializing an ipc header.