Implement IPC_TYPE_GET_MONITORS message request
These functions use yajl to generate a JSON document detailing the properties
of each of the monitors and their clients. This message is sent back to the
requesting client. The JSON document contains info effectively about the entire
window manager layout and how clients are arranged.

The ystr macro significantly simplifies the yajl_gen_string function and avoids
unnecessary repetition.