Correct dump_monitors properties
The arrays of two elements in the monitor structs are arrays of current and old
states which are indicated by its sellt, seltags properties. These sel*
properties hold the index of the current layout and tagset while the remaining
index points to the old layout and tagset.

Also, dumping all of the client information in this request results in a very
large message with over 4000 bytes for just a few windows with the JSON document
pretified. Instead, list clients by their window id and implement another
message type for getting a particular client's properties.