Implement IPC_TYPE_GET_CLIENT
Modify dwm.c to handle IPC_TYPE_GET_CLIENT message type. Add a res variable to
check the validity of the client window id before trying to dump its properties.

Add ipc_parse_get_client to parse an IPC_TYPE_GET_CLIENT message and get a
pointer to the client referred to be the client window id that should be
specified in the client's request message. This function returns -1 if the
client does not exist.

Add ipc_get_client to generate a JSON response with the specified dwm client's
properties to be sent back to the IPC client.