Check if fd is registered before handling event
This was to deal with a side-effect of the previous issue where passing a new
epoll_event with an uninitialized data member would cause epoll to start
monitoring a random file descriptor.

This is an additional check to make sure the fd is registered before attempting
to handle any events on it. Theoretically, this check should never fail as long
as fds that are registered with epoll are clients that actually requested a
connection. A case as described above could cause the check to fail, but this
check would prevent a crash and instead exit more gracefully.