Integrate ipc-related epoll commands into ipc.c
The main problem with adding client fds to epoll and updating their events in
dwm.c was being unable to cleanly add the EPOLLOUT event for sending events to
subscribers. Since the event would be triggered from a dwm function thats
related to the event, there would have to be some logic to add EPOLLOUT to the
clients that should receive the event message. Adding this in dwm.c would
require iterating through all the clients with pending buffers and then setting
adding the EPOLLOUT event, but that would also require knowing the prior
epoll_event.events setting to avoid removing previously set EPOLL* events.

The other option would be to add the logic in ipc.c, but then ipc.c would be
manipulating epoll and it wouldn't make sense to have client/socket epoll
commands existing in both dwm.c and ipc.c. As a result, all the socket/client
epoll related commands were integrated into ipc.c, so as messages are prepared
EPOLLOUT is automatically added as an event for epoll to monitor. When clients
are added/disconnected, ipc.c handles the epoll commands.

The main issues with integrating epoll into ipc.c is that now it will be a bit
messier if kqueue or other epoll-like system libraries are implemented for
other operating systems. If all epoll logic was in dwm.c, it may have been
easier to implement other epoll-like libraries as needed to increase
portability. In addition, now there are two different places from which epoll is
being manipulated which could increase the obscurity of how epoll is manipulated
in the future.

Nonetheless, the best and cleanest option at this moment is to integrate epoll
into ipc.c.

ipc_init will now take the epoll fd as an argument for future epoll manipulation
in ipc.c and ipc_create_socket will now be an internal function that is called
by ipc_init.