Implement IPC_TYPE_SUBSCRIBE/IPC_TYPE_EVENT
At the moment only the tag_change_event has been implemented. This was difficult
to implement since there was no stored history of the previous occupied/urgent
tags. The occupied/urgent tags were determined in drawbar() in dwm.c and not
stored as a member of the Monitor struct as the tagset is. Additional arrays
could have been added as occupied[2] and urgent[2] in a similar manner to
tagset[2], but this would require adding additional members to keep track of
which element is active and more significantly require additional code in
multiple places to update these members. This method may be looked into in the
near future as it may be cleaner and take much less code. The main issue with
this method is that it would require more modifications to dwm.c and decrease
compatability with other patches.

As a result, a IPCTagState struct was created to keep track of the old tag state
of each monitor which came with additional code for manipulating this linked
list. In drawbar(), the update_tag_state function is called which checks if a
tag change has occured and notifies subscribers accordingly.

Some helper functions were added for creating new IPCTagState objects and
manipulating the linked list.

dump_tag_state and dump_tag_event functions generate the main JSON for the event
message.

ipc_event_stoi converts the event string to its enum value.

ipc_parse_subscribe parses an IPC_TYPE_SUBSCRIBE message and determines if a
client is trying to subscribe or unsubscribe and which event it is referring to.

ipc_subscribe subscribes/unsubscribes a given client from a specified event.

ipc_update_tag_state checks if the selected, occupied, and urgent tags have
changed and calls ipc_tag_event if they have, passing the old and new tag
states.

ipc_tag_event generates a JSON message with the old and new tag states and
prepares a message to send to subscribed clients.