Refactor tag_change_event
Storing TagStates in ipc.c as a linked list is very unclean and introduces a lot
of extra code for manipulating the linked list, allocating memory, etc.
This commit improves the method of detecting and sending this event.

First, the IPCTagState (now TagState) type is stored in types.h which is a
better place for this type to be defined, semantically speaking.

The Monitor struct now contains a TagState struct embedded which contains the
last known values of the occuped, urgent, and selected tags. After any XEvent
occurs, the current occupied, urgent, and selected tags can be calculated and
compared to the TagState member of the Monitor struct for every monitor. If the
structs are different, this indicates a tag_change_event which then calls the
ipc_tag_change_event.

This method of calling the event separates the tag change event from the drawbar
function where it didn't really belong in the first place. Since this check
occurs after every XEvent and after an IPC_TYPE_COMMAND_RUN, (drawbar also
happens almost after every XEvent), it would be unlikely for a tag change to
occur without being detected. Also, since it is separate from drawbar, and since
the code change to dwm.c is minimal (and mostly additive), this method does not
signficantly impact the compatability of this patch with other patches.

The monitor number is now also not part of the TagState struct and instead if
passed along as part of the tag_change_event as a dictionary entry on the same
level as old_state and new_state. This makes more sense, since the monitor
number will always be the same for state changes.