Implement IPC_EVENT_SELECTED_CLIENT_CHANGED
updatelastsel in dwm.c calls the ipc.c ipc_selected_client_changed_event
function if the selected client on any of the monitors changed which sends a
message to subscribers with the properties of the old client and the new client
and the monitor number for which the selection change happened.

Also add lastsel member to types.h to keep track of last known selected client.