Remove extra layer of indirection for run command
The parse_run_command function took an argument of Arg(**)[] which was a pointer
to a pointer to an array of Arg. There was no reason for the extra indirection
which would allocate heap memory for a pointer to each of the Arg structs.
Instead, using Arg(*)[], one contigous block of memory can be allocated where
the structs lie beside each other. This avoids extra memory being allocated,
reduces the chance of memory leaks, and simplifies the code.