Refactor IPC_TYPE_RUN_COMMAND message handling
Instead of hardcoding the IPC run commands, define the available commands in
config.def.h. The IPCCommand struct is created to associate the command name
with a pointer to the function and the number of arguments the function/command
takes.

First, this allows easy extensibility of IPC commands by easily being
able to add them to the list in config.def.h and then simply define them in
dwm.c. This also allows functions with multiple arguments to be used using the
ArgFunction union which can hold a simple void function pointer that takes a
const Arg* or a void function pointer that takes a const Arg* and an additional
number of arguments. Since most of the standard dwm functions only take one
argument, this allows compatability with dwm's functions as well as
expandability to define your own that take multiple arguments.

The IPCCommand struct also eliminates the need to associate each command with a
function using if/switch statements. The IPC command enum and ipc_command_stoi
since it simply added extra code that wasn't necessary. Also, it was a kind of
dirty way to make a hash map.

Since dwm.c has no header file, the only clean way to reference the functions in
dwm.c and make ipc.c know of them is to pass a static array of these structs to
the ipc_init function which then stores the address to this array and the length
of it for later reference.

The ipc_run_command function takes the name of the command as an argument as
well as an array of Arg structs and the number of arguments. It looks for a
matching command name and then executes the correct function based on whether it
takes a single argument or multiple arguments (in which case the number of
arguments is also passed). This function also does a simple check to see if the
number of arguments provided match the number of arguments the function takes.

This commit ultimately creates a cleaner way to define IPC commands and reduces
code in dwm.c.