Refactor ipc run command: free strings, type check
Add a new member arg_types to the IPCCommand struct which holds the address to
an array of the expected arugment types in order. This pointer should be
assigned an address to a static array. This allows type checking before trying
to run commands with arbitrary arguments while argc checks the number of
arguments.

The IPCCOMMAND macro enables an elegant and simple way of defining IPC commands
in config.h where the first argument is the name of the function to call in
dwm.c (also the name of the command), the second argument is the number of
Arg this function expects, and the third argument is an order-sensitive array of
types specified by the ArgType enum.

Since the Arg union can hold a void pointer, the ARG_TYPE_STR specifies that a
string should be passed to the corresponding ipc command function whose address
would be stored in Arg->v. The string would be allocated during
ipc_parse_run_command and should be freed accordingly after calling the
corresponding function if the ARG_TYPE_STR type matches.

If an actual pointer address is passed, this address should be passed as a
number in the JSON message and will be casted to a void pointer and stored in
Arg->v.

The ipc_get_ipc_command function gives an IPCCommand given the name of the
command.

ipc_parse_run_command now does command name checking, type checking, and number
of arguments checking. This is done in the parsing function, so the args
returned match expected args and their types, and so ARG_TYPE_STR arguments can
be identified and freed after their use. This function needs to be refactored
and made prettier if possible.

The ipc_get_command function no longer does any argument checking since it is
now handled in ipc_parse_run_command. Instead, the function simply calls the
corresponding functions if the message is parsed correctly and then frees any
ARG_TYPE_STR arguments.