Move find client code to ipc_get_dwm_client
To keep things more consistent and cleaner, just loop through the monitors and
clients to find the client of the given Window in ipc_get_dwm_client rather than
doing it in dwm.c. Also call the parse the client in ipc_get_dwm_client.

This keeps most of the handling code for this message type to ipc.c.