Add dwm-msg program
dwm-msg is a cli tool for sending messages to dwm through the socket. It
supports all of the message types that are currently implemented and can be used
to retreive properties, run commands, and subscribe to events.

yajl is used for generating the JSON messages because it is a lot easier for
more complicated messages and has good error-checking.

Update the Makefile to compile dwm-msg.

Add dwm-msg to .gitignore.