Include source files in dwm.c
Since the VERSION symbol is defined when being compiled, only include the
types.h file when VERISON is not defined. This allows the types.h types to be
defined in dwm.c which signficantly improves patch compatability. Since these
types would not be removed from dwm.c, the final patch has minimal deletions in
dwm.c

Since the types are now defined in dwm.c only when VERSION is defined, all of
the type definitions in dwm.c are included in types.h, to avoid missing
definition issues when running clang.

Since there are no cyclic includes with yajl_dumps and IPCClient, these source
files can be directly included in dwm.c as well and don't have to be compiled
separately.

Overall this commit improves the compatbility of the patch in a bit of a hacky
way by minimizing the deletions in dwm.c