Close sockets properly
Use shutdown to stop communication and then close to actually close the socket.

shutdown does not close the file descriptor, but is only used to mark the end of
further communication. close actually closes the file descriptor. Without this,
file descriptors could be kept open confusing clients and making it difficult
for clients to tell if the socket has been closed.