Implement focused_title_change_event
This event is triggered when a selected client's title changes. This event is
raised directly from the dwm function updatetitle since it would be the most
performant and since this function is unlikely to be modified by most patches
(compatbility shouldn't be hurt much, only additions). The alternative would
require adding another member variable to hold the client name and comparing it
every X event.

dwm-msg.c is also updated to support this event.

This commit was made in support of the following issue:
https://github.com/mihirlad55/polybar-dwm-module/issues/4