Fix segfault from dump_monitor
This was being caused by dumping a monitor when it had no selected clients.
There was no check to see if Monitor->sel was a null pointer or a valid client,
so accessing the Monitor->sel->win member would segfault. This in most cases
would not happen, but at startup when no windows are selected, this would
consistently cause a crash.