Implement focused_state_change_event
This event is raised when the currently selected client on any monitor has a
state change i.e. if isfullscreen, isfixed, etc. changes.

To keep monitoring these states simple, a tagstate-like approach is used to keep
track of the previous state and compare the previous state with the current
state.