dwm-msg: Add --ignore-reply option
This option ignores the initial result of sending the run_command or subscribe
command. This is useful for scripting, particularly with the subscribe command,
where you know the result will be a success and you don't want to have to deal
with the initial result message of subscribing, but rather the events
themselves.