config.def.h: Remove spawn from ipccommands
This command is basically useless since there is no way to retrieve the address
of a char array that defines a command to spawn. Plus even if there was a way,
there's no reason to use an IPC command to execute a command, since this can
easily be done using bash or popen, etc.