Fix ARG_TYPE_NONE passing non-zero value
Since ARG_TYPE_NONE would treat the union as a float when setting it to 0,
only 4 bytes of it would be set to 0, leading to uninitialized data when
interpreting the union as other data types such as a long (8 bytes).

Treating it as a long when setting it to 0 should fix this issue.

This fixes issue #14.